Performance improvements and other changes.
